JAMES-1818 Move Cid and MessageAttachment to API
